modified

+++ Models/config.go
@@ -8,7 +8,7 @@ var err error
 
 type ModelInterface interface {
 	GetArticle() ([]*Article, error)
-	GetArticleById(id uint) (*Article, error)
+	GetArticleById(id uint64) (*Article, error)
 	CreateArticle(article *Article) error 
 	UpdateArticle(article *Article) error
 	DeleteArticle(id uint64) error
